+++ src/models/interfaces.ts
@@ -27,6 +27,8 @@ export interface IElementStyle {
   foregroundHex: string;
   inactiveBackgroundHex: string;
   inactiveForegroundHex: string;
+  badgeBackgroundHex?: string;
+  badgeForegroundHex?: string;
 }
 
 export interface IPeacockAffectedElementSettings {